==> scripts/01 - Data collection.ipynb <==
# Modules and scripts <a class="anchor" id="chapter1"></a>

import requests
from bs4 import BeautifulSoup
import pandas as pd
import os
import re
import pickle
import scraping_functions

# Scraping the film IDs and the URLs <a class="anchor" id="chapter2"></a>

IMDB doesn't have an API. However, there are two unofficial IMDBI APIs: The Open Movie Database (OMDb) and The Movie Database (TMDB). In this project we'll rely on the latter as it doesn't enforce a rate limit. We're also going to supplement the API data with scraped data.

TMDB API works by supplying the ID of the film(s) whose info we'd like to retrieve. The IDs consist of a double "t" followed by seven or eight digits. They can be found inside a film URL. For example, "tt0050782" is the ID for this film: https://www.imdb.com/title/tt0050782/. 

The first step will thus be retrieving the IDs of the 5 000 most popular films on IMDB, which can be found in this [playlist](https://www.imdb.com/search/keyword/?mode=detail&page=1&title_type=movie&ref_=kw_ref_rt_vt&num_votes=5000%2C&sort=num_votes,desc).

To scrape the film IDs of our 5 000 films we'll iterate a for loop over the first 100 pages in the playlist (each containing 50 films). The film ID is the value of the 'href' attribute inside h3 elements with the "lister-item-header" class. We'll write a simple regex to extract the film ID. We'll also scrape the entire URL: it's going to come in handy later when we do the scraping.

film_ids = []
film_urls = []

for i in range(1, 101):
  
  content = requests.get(f"https://www.imdb.com/search/keyword/?mode=detail&page={i}&title_type=movie&ref_=kw_ref_rt_vt&num_votes=5000%2C&sort=num_votes,desc").content
  
  soup = BeautifulSoup(content)

  for film in soup.find_all("h3", class_="lister-item-header"):

    link = film.find("a").get("href")
    film_id = re.findall(pattern = r"tt\d+", string = link)[0]
    film_ids.append(film_id)

    film_url = "https://www.imdb.com" + link
    film_urls.append(film_url)
    

Saving both the film IDs and the URLs

with open('data/film_ids.txt', 'wb') as f:
    pickle.dump(film_ids, f)
    
with open('data/film_urls.txt', 'wb') as f:
    pickle.dump(film_urls, f)

# API calls <a class="anchor" id="chapter3"></a>

Let's build the dataframe with the API data by feeding the film IDs to our function

df = scraping_functions.build_film_df(film_ids)
df.head()

df.shape

Let's save it as a .csv file

df.to_csv("data/df_api.csv")

# Scraping additional data <a class="anchor" id="chapter4"></a>

TMDB API has some limits: it doesn't provide data about the people who worked on a film (directors, writer, actors etc.) and its rating data is inaccurate. We'll thus integrate the data we just got by scraping some more information.

To build the dataframe of scraped data we'll import another function from our functions scripts and we'll run it on the film URLs we scraped in section 2

df2 = scraping_functions.build_scraped_df(film_urls)

Let's have a look at the dataframe

df2.head()

Let's save the dataframe

df2.to_csv("data/scraped_df_1.csv")

# Scraping people data <a class="anchor" id="chapter5"></a>

As mentioned previously, TMDB API doesn't provide much data about the people who worked on a film, so we'll need to scrape that data ourselves. These pieces of info aren't located in the regular film pages but in the "Full cast and crew" pages. So before scraping the data we first need to collect all these pages.

with open('data/film_ids.txt', 'rb') as f:
    film_ids = pickle.load(f)

crew_pages = []

for i in range(len(film_ids)):
    crew_page = f'https://www.imdb.com/title/{film_ids[i]}/fullcredits/?ref_=tt_cl_sm'
    crew_pages.append(crew_page) 

## Creating the dataframe <a class="anchor" id="subparagraph4"></a>

Let's run it on our list of 'Full cast and crew' pages

df3 = scraping_functions.build_scraped_df_2(crew_pages)

Let's have a look at the dataframe

df3.head()

Let's save it

df3.to_csv('data/scraped_df_2.csv')
