==> scripts/02 - EDA and data cleaning.ipynb <==
# Modules and data

import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import re
import os
import pickle
import scraping_functions # Let's also import the script with our scraping functions

Loading the first dataframe (i.e. the one with the API data)

df = pd.read_csv("https://raw.githubusercontent.com/JacopoMalatesta/imdb_most_popular_films/main/data/df_api.csv",
              index_col = 0,
              dtype = 'object')

Loading the second dataframe (i.e. the first dataframe of scraped data)

df2 = pd.read_csv("https://raw.githubusercontent.com/JacopoMalatesta/imdb_most_popular_films/main/data/scraped_df_1.csv", 
                  index_col = 0,
                  dtype = 'object')

Loading the third dataframe (i.e. the second dataframe of scraped data)

df3 = pd.read_csv("https://raw.githubusercontent.com/JacopoMalatesta/imdb_most_popular_films/main/data/scraped_df_2.csv",
                  index_col = 0,
                  dtype = 'object')

Let's also load the film IDs

with open('data/film_ids.txt', 'rb') as f:
    film_ids = pickle.load(f)

# First datraframe

df.head()

All IDs are unique

df["id"].duplicated().sum()

We have virtually zero null values

df.apply(lambda col: col.isnull().sum() / df.shape[0])

# Second dataframe

Let's now have a look at second dataframe. This is the first of two datasets containing scraped data.

df2.head()

Same number of rows as the dataframe of API data

df.shape[0] == df2.shape[0]

There's one duplicate ID.

df2["id"].duplicated().sum()

This is because we have two null values in the ID column

df2["id"].isnull().sum()

11% of the rows have missing values in the color variable, 7% in the metascore variable

df2.apply(lambda col: col.isnull().sum() / df2.shape[0]).sort_values(ascending = False)

Are these missing values the result of some mistake we made while scraping or are they actually missing on IMDB?

The color information is missing from the IMDB pages of these films. So the NAs in this Series are not due to mistakes in the scraping process.

df2[df2["color"].isnull()]

Same thing for the 'metascore' Series: the Metascore of the films with missing values in this column is also missing on IMDB.

df2[df2["metascore"].isnull()]

Same thing for the 'aspect_ratio' Series

df2[df2['aspect_ratio'].isnull()]

## Filling missing values

### Missing films

For some reason we couldn't scrape any info about seven films.

df2[df2['director'].isnull()]

Let's extract the film IDs of the three missing films into a Pandas Series

missing_film_ids = list(df2[df2['director'].isnull()]['id'])

Let's fill the two missing IDs

missing_film_ids[1] = film_ids[991]
missing_film_ids[4] = film_ids[3162]

Let's fill them in the dataframe as well

df2.loc[991, 'id'] = film_ids[991]
df2.loc[3162, 'id'] = film_ids[3162]

Let's scrape those pieces of data again 

for i in range(len(missing_film_ids)):
    
    content = requests.get(f"https://www.imdb.com/title/{missing_film_ids[i]}").content
    soup = BeautifulSoup(content)
    
    df2.loc[df2['id'] == missing_film_ids[i], 'director'] = scraping_functions.scrape_director(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'writer'] = scraping_functions.scrape_writer(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'imdb_rating'] = scraping_functions.scrape_imdb_rating(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'imdb_rating_count'] = scraping_functions.scrape_rating_count(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'metascore'] = scraping_functions.scrape_metascore(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'user_review_count'] = scraping_functions.scrape_user_review_count(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'critic_review_count'] = scraping_functions.scrape_critic_review_count(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'color'] = scraping_functions.scrape_color(soup)
    df2.loc[df2['id'] == missing_film_ids[i], 'aspect_ratio'] = scraping_functions.scrape_aspect_ratio(soup)

We successfully scraped the data for all films.

df2[df2['id'].isin(missing_film_ids)]

### Missing user review counts

df2[df2['user_review_count'].isnull()]

Let's extract the IDs of these films in a Pandas series

missing_user_review_ids = df2.loc[df2["user_review_count"].isnull(), "id"]

Let's scrape the user review count again

for i in range(len(missing_user_review_ids)):
    content = requests.get(f"https://www.imdb.com/title/{missing_user_review_ids.iloc[i]}").content
    soup = BeautifulSoup(content)
    
    df2.loc[df2['id'] == missing_user_review_ids.iloc[i], 'user_review_count'] = scraping_functions.scrape_user_review_count(soup)

We managed to retrieve the user review count for all those films.

df2[df2["id"].isin(missing_user_review_ids)]

These are the % of null values now

df2.apply(lambda x: x.isnull().sum() / df2.shape[0]).sort_values(ascending=False)

# Third dataframe

df3.apply(lambda x: x.isnull().sum() / df3.shape[0]).sort_values(ascending=False)

No duplicate IDs

df['id'].duplicated().sum()

Films with missing 'cinematographer' are mostly animated films (which do not have a cinematographer)

df3[df3['cinematographer'].isnull()]

# Joining the dataframes

Let's join the second dataframe to the first one by using the film id as our key. They have exactly the same IDs so a left, inner and right join would all result in the same result.

cleaned_df = pd.merge(left = df, right = df2, how = 'inner', on = 'id')

Let's now join the third dataframe to the other two.

cleaned_df = pd.merge(left = cleaned_df, right = df3, how = 'inner', on = 'id')

Let's have a look at the output

cleaned_df.head()

We ended up with the right number of rows

cleaned_df.shape

cleaned_df.columns

cleaned_df = cleaned_df.reindex(columns= ['id', 'title', 'release_date', 'runtime', 'country', 'language', 'genre', 'studios',
                             'color', 'aspect_ratio', 'budget', 'revenue', 'imdb_rating', 'imdb_rating_count', 
                             'metascore', 'user_review_count', 'critic_review_count', 'director', 'writer', 
                             'actors', 'cinematographer', 'editor', 'composer', 'production_designer', 'art_director',
                             'costume_designer', 'producers', 'last_updated'])

# Casting

Right now all the columns are of the 'object' type. We need to cast them to the correct dtype.

cleaned_df.info()

Let's first cast the following columns to integers. The ones with missing values will actually be cast to floats.

integer_cols = ['runtime', 'budget', 'revenue', 'imdb_rating_count', 'metascore', 'user_review_count', 'critic_review_count']

cleaned_df[integer_cols] = cleaned_df[integer_cols].apply(lambda x: pd.to_numeric(arg = x, downcast='integer'))

Let's cast the release date to a datetime variable

cleaned_df[['release_date', 'last_updated']] = cleaned_df[['release_date', 'last_updated']].apply(pd.to_datetime)

Let's cast the IMDB avg rating to float

cleaned_df['imdb_rating'] = pd.to_numeric(cleaned_df['imdb_rating'])

Everything looks fine

cleaned_df.info()

# Saving the cleaned dataframes

cleaned_df.to_csv("data/cleaned_df.csv")
